requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking pyproject.toml
#
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.5
    # via typer
google-api-core==2.11.1
    # via google-api-python-client
google-api-python-client==2.86.0
    # via kevinfjiang (pyproject.toml)
google-auth==2.22.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.59.1
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
mdutils==1.6.0
    # via kevinfjiang (pyproject.toml)
protobuf==4.23.4
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pyparsing==3.1.0
    # via httplib2
pyyaml==6.0
    # via kevinfjiang (pyproject.toml)
requests==2.31.0
    # via google-api-core
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
typer==0.9.0
    # via kevinfjiang (pyproject.toml)
typing-extensions==4.7.1
    # via typer
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
